tee logging to logrus instead of discarding the logs locally
When logrus logging works, great! However, we often want to run and
debug locally. Howveer, we have nice run-scripts which divert all
logging to appropriate local files, and we don't want to discard that;
it can be faster and simpler to just grep a file than to get into
the whole Honeycomb thing.